bridge/requirements.txt
paho-mqtt==2.1.0
requests==2.31.0
python-dotenv==1.0.1
Flask==3.0.0
flasgger==0.9.7.1

# ML dependencies for bird classification
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=10.0.0

# European bird classifier (birder-project)
birder>=0.1.0
